# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Login + ToDo API",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "Методы для регистрации, входа и получения профиля пользователя"
    },
    {
      "name": "Todo",
      "description": "Методы для управления списком задач пользователя"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Регистрация пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная регистрация"
          },
          "400": {
            "description": "Пользователь уже существует"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход и токен"
          },
          "400": {
            "description": "Неверный логин или пароль"
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Получение профиля",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный доступ"
          },
          "401": {
            "description": "Нет токена"
          },
          "403": {
            "description": "Неверный токен"
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": [
          "Todo"
        ],
        "summary": "Получить список задач пользователя",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список задач пользователя"
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "summary": "Добавить новую задачу",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  }
                },
                "required": [
                  "text"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Задача добавлена"
          },
          "400": {
            "description": "Ошибка создания задачи"
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "tags": [
          "Todo"
        ],
        "summary": "Редактировать текст задачи",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  }
                },
                "required": [
                  "text"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Задача обновлена"
          },
          "404": {
            "description": "Задача не найдена"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "summary": "Удалить задачу",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Задача удалена"
          },
          "404": {
            "description": "Задача не найдена"
          }
        }
      }
    },
    "/todos/{id}/status": {
      "patch": {
        "tags": [
          "Todo"
        ],
        "summary": "Изменить статус задачи",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completed": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "completed"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус обновлён"
          },
          "404": {
            "description": "Задача не найдена"
          }
        }
      }
    }
  }
}